move manager composables to manager folder
