+++ src/workbench/extensions/manager/composables/nodePack/usePacksSelection.ts
