+++ src/workbench/extensions/manager/composables/nodePack/useUpdateAvailableNodes.ts
@@ -1,8 +1,8 @@
 import { compare, valid } from 'semver'
 import { computed, onMounted } from 'vue'
 
-import { useInstalledPacks } from '@/composables/nodePack/useInstalledPacks'
 import type { components } from '@/types/comfyRegistryTypes'
+import { useInstalledPacks } from '@/workbench/extensions/manager/composables/nodePack/useInstalledPacks'
 import { useComfyManagerStore } from '@/workbench/extensions/manager/stores/comfyManagerStore'
 
 /**
