+++ src/workbench/extensions/manager/composables/useRegistrySearch.ts
