+++ src/workbench/extensions/manager/stores/conflictDetectionStore.ts
